requirements.txt

Flask==3.0.0
Pillow==10.2.0
Flask-PyMongo==2.3.0
requests==2.31.0
